Move pdiff.diff to textual-diff
  ref #18
